added crx helpers to core
