+++ packages/altair-core/package.json
@@ -36,8 +36,10 @@
     "@jedmao/location": "^3.0.0",
     "@jest/globals": "^29.7.0",
     "@types/actioncable": "^5.2.5",
+    "@types/chrome": "^0.0.209",
     "@types/color-name": "^1.1.4",
     "@types/crypto-js": "^4.0.1",
+    "@types/firefox-webext-browser": "^109.0.0",
     "@types/jest": "^29.5.12",
     "@types/json-bigint": "^1.0.1",
     "@types/node": "^14.14.41",
