+++ packages/altair-crx/package.json
@@ -30,6 +30,7 @@
   "devDependencies": {
     "@types/adm-zip": "^0.5.7",
     "@types/chrome": "^0.0.209",
+    "@types/firefox-webext-browser": "^109.0.0",
     "@types/react": "^18.0.0",
     "@types/react-dom": "^18.0.0",
     "@vitejs/plugin-react": "^1.3.0",
